requirements.txt
# Core dependencies
redis>=4.5.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# AI and ML dependencies
openai>=1.0.0
anthropic>=0.7.0
tiktoken>=0.5.0

# Web framework and API
flask>=2.3.0
flask-cors>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Async and concurrency
asyncio-mqtt>=0.13.0
aioredis>=2.0.0
asyncio>=3.4.3

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.0.0
python-json-logger>=2.0.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
toml>=0.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Docker and deployment
gunicorn>=21.0.0
docker>=6.1.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0

# Database (for state persistence)
sqlalchemy>=2.0.0
alembic>=1.11.0

# Utilities
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0
requests>=2.31.0

